sled agent: pick up Propolis migration status changes
Pick up a Propolis commit containing propolis#661. This changes Propolis's
migration monitoring API to report migrations in and migrations out as
independent objects.

Rework the sled agent logic that processes incoming Propolis notifications to
consume this information. The new logic is not all that much different from the
old, except that the consequences of observing a mismatched migration ID are
slightly less ambiguous now: in the old code, this was interpreted as "well,
maybe there's a new migration that Propolis hasn't found out about yet"; in the
new code, this is more clearly an unexpected condition, albeit one sled agent
handles in the same way as before (by punting and updating the relevant VMM
state while leaving the instance state alone).

Tests: cargo test in omicron-nexus and omicron-sled-agent.